Reset returns initial state, state format fix
Reset now returns initial state of the environment so that it may be
used immidiately.

Environment state now follows proper format, four discrete numbers in a
single tuple (as opposed to a tuple consisting of two tuples).

To avoid repetition the _get_obs function is added, which returns the
current state of the environment.